refactor(lex-tables): increase lex table state size
I want to add more handlers, but to do this, I need more bits in the
state. Add more bits by widening the state type from 8 bits to 16 bits.

In the future, I plan to reduce the number of data bits from 5 to 4, in
which case we can undo this 8->16 change.

+++ src/quick-lint-js/fe/lex-tables.h
@@ -179,7 +179,8 @@ struct lex_tables {
   static_assert(character_class_table[static_cast<std::uint8_t>(u8'|')] == character_class::pipe);
   // clang-format on
 
-  using state_type = std::uint8_t;
+  // TODO(strager): Shrink to 8 bits.
+  using state_type = std::uint16_t;
 
   // How many bits in the state are reserved for the state number for
   // intermediate and non-unique terminal states, or for extra data for unique
